Add tasks to patch the UFW systemd unit file for Kali linux
The UFW systemd unit file for Kali linux contains a
Wants=network-pre.target that is unnecessary.  It also lacks a
DefaultDependencies=no line that is necessary.  These errors cause a
dependency cycle in systemd, which systemd breaks by refusing to start
the cloud-init service even if it is enabled.  This means that the
instance userdata is not run, so this is a serious problem.